% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset-documentation.R
\docType{data}
\name{brazil}
\alias{brazil}
\title{Parts of the ?YEAR? census in Brazil}
\format{A data frame with ?? rows and ?? variables:
\describe{
 \item{geo2_br}{Consistent level two boundaries of Brazil, Municipality 1980 - 2010}
 \item{age}{The age of a person in years}
}}
\source{
\url{https://international.ipums.org/international/index.shtml}
}
\usage{
data(brazil)
}
\description{
The dataset contains ?NUMBER? individual observations from the ?YEAR? census in Brazil. 
Number of variables and observations are reduced to a size for a reasonable example 
of our ELLsae package
}
\keyword{datasets}
