\name{ELLsae-package}
\alias{ELLsae-package}
\alias{ELLsae}
\docType{package}
\title{
\packageTitle{ELLsae}
}
\description{
\packageDescription{ELLsae}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ELLsae}
\packageIndices{ELLsae}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ELLsae}

Maintainer: \packageMaintainer{ELLsae}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
