% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELLsae_big.R
\name{ELLsae_big}
\alias{ELLsae_big}
\title{ELLsae_base}
\usage{
ELLsae_big(model, surveydata, censusdata, location_survey, location_census,
  mResponse, n_boot = 50, seed, welfare.function, transf, transf_inv,
  output = "default", num_cores = 1, quantiles = c(0, 0.25, 0.5,
  0.75, 1), save_boot = F)
}
\arguments{
\item{model}{a model that is specified for the relationship betwenn the 
response varibale and the regressors. Model must be a linear model that can 
be processed by \code{lm()}}

\item{surveydata}{Smaller surveydata with additional response variable of 
interest. Will be used to estimate the linear model}

\item{censusdata}{The dataset in which the response from \code{model} is 
supposed to be imputed}

\item{location_survey}{Name of location variable or vector for the survey 
data which is used for error correction and the location means (if 
\code{mResponse} is specified)}

\item{location_census}{name of location variable (string) in the census data 
which is used for error correction and location means. If \code{mResponse} 
is specified, but \code{location_census} is missing}

\item{mResponse}{Additional parameters for the regression based on location 
means calculated from the census data to account for the lack of 
information in a small survey}

\item{n_boot}{Number of bootstrap samples used for the estimation, default is 
\code{n_boot = 50}}

\item{welfare.function}{Additionally a welfare function for the response can 
be specified}

\item{output}{indicator for which output is requested as a list.}

\item{parallel}{indicates if compution is supposed to be done in parallel to 
improve speed}

\item{save_yboot}{logical indicator if the bootraps of the response y are 
supposed to be saved as a CSV file under your current working direktory. 
The name is: ...}
}
\value{
The function takes the the typically smaller surveydata and uses the 
argument \code{model} to estimate a linear model of the type \code{lm()}. In case
the argument \code{mResponse} is specified means from the cluster data for the given 
variables are calculated and merged with the survey databy cluster locations. These
new explanatory variables are also used for the estimation of the linear model. 

In the second step a C++ fuction takes over and calculates \code{nboot} predicted 
Y´s by using the betas from the first step to draw from a multivariate normal distribution
and draws indicidual and nested errors at random with replacement. If requested the Y´s 
are used to estimate the welfare function after which either the mean of the yhat or of the 
welfare fuctiom is returned.

The function returns a list with different objects. If \code{output} 
is left unspecified the estimated Y´s or welfare estimates \code{yhat}, 
the fit of the linear model \code{model_fit} and a summary of the bootstrap samples
\code{bootstrapCI} are returned.

If \code{output} is specified all the arguments given are returned. Next to the 
above the following inputs are possible, \code{surveydata} and \code{censusdata}
return the data frames used for the compution as some rows might be deleted due 
to NA handling and additional variables are created for \code{mResponse}. 

Additionally the bootstrapped Y´s can be saved as a CSV if \code{save_yboot} is 
set equal \code{TRUE} and can be found under the current working directory as 
"Bootraps-of-Y.csv".
}
\description{
\code{ELLsae_base} is a method for small area estimation used to
 impute a missing 
variable from a smaller survey dataset into a census. The imputation is based 
on a linear model and bootstrap samples.
}
\examples{
mean(c(1,2,3,4))
}
\references{
Elbers, C., Lanjouw, J. O. and Lanjouw, P. (2003). \emph{Micro-Level Estimation of Poverty and Inequality}. 
  In: Econometrica 71.1, pp. 355-364, Jan 2003

  Guadarrama Sanz, M., Molina, I., and Rao, J.N.K.  (2016). \emph{A comparison of small 
  area estimation methods for poverty mapping}. In: 17 (Mar. 2016), 41-66 and 156 and 158.
}
\seealso{
Other small area estimation methods can also be found 
in the package \code{sae}.
}
\keyword{SAE,}
\keyword{imputation}
