% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELLsaeBig.R
\name{ELLsaeBig}
\alias{ELLsaeBig}
\title{ELLsaeBig}
\usage{
ELLsaeBig(model, surveydata, censusdata, location_survey, mResponse,
  location_census, n_boot = 50, welfare.function, test, parallel = F,
  output, save_yboot = F)
}
\arguments{
\item{model}{a model that is specified for the relationship betwenn
the response varibale and the regressors. Model must be a linear model that can be processed by \code{lm()}}

\item{surveydata}{Smaller surveydata with additional response variable of interest.
Will be used to estimate the linear model}

\item{censusdata}{The dataset in which a certain variable is supposed to be imputed}

\item{location_survey}{Name of location variable or vector for the survey data which is used for
error correction and the location means (if \code{mResponse} is specified)}

\item{mResponse}{Additional parameters for the regression based on location means
calculated from the census data to account for the lack of information in a small survey}

\item{location_census}{name of location variable (string) in the census data 
which is used for error correction and location means. If \code{mResponse} 
is specified, but \code{location_census} is missing}

\item{n_boot}{Number of bootstrap samples used for the estimation, default is 50}

\item{welfare.function}{Additionally a welfare function for the response can be specified}

\item{parallel}{indicates if compution is supposed to be done in parallel to 
improve speed}

\item{output}{indicator for which output is requested as a list.}

\item{save_yboot}{logical indicator if the bootraps of the response y are 
supposed to be saved as a CSV file under your current working direktory. 
The name is: ...}
}
\value{
Was die Funktion ausspuckt.
}
\description{
Beschreibung der Funktion
}
\examples{
1-2
}
\references{
Elbers, C., Lanjouw, J. O. and Lanjouw, P. (2003). \emph{Micro-Level Estimation of Poverty and Inequality}. 
  In: Econometrica 71.1, pp. 355-364, Jan 2003
   
  Guadarrama Sanz, M., Molina, I., and Rao, J.N.K.  (2016). \emph{A comparison of small 
  area estimation methods for poverty mapping}. In: 17 (Mar. 2016), 41-66 and 156 and 158.
}
\seealso{
Other small area estimation methods can also be found in the package \code{sae}.
}
\keyword{SAE}
\keyword{ellsea,}
