% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomSAEdata.R
\name{generateSAEdata}
\alias{generateSAEdata}
\title{generateSAEdata}
\usage{
generateSAEdata(nObs = 30, nAreas = 50, betas = c(25, -0.029, 45.5),
  fraction = 0.2)
}
\arguments{
\item{nObs}{number of observations per area}

\item{nAreas}{the number of areas, default is 100}

\item{betas}{vector of three containing betas for the estimation of the response}

\item{fraction}{split between survey and census data (between 0 and 1)}
}
\value{
Generates a list containing random \code{surveydata} and random \code{censusdata}. The variables
\code{x1, x2} are generated from bernoulli distribitions and \code{x2} is generated by a poisson distribution
with lambda equal to the location. Next to an individual error a nested error is used to create the response.
}
\description{
Provides example data for ELLsea functions and returns 
a survey and census dataset
}
\section{Functions}{
\itemize{
\item \code{generateSAEdata}: example data for ELLsea functions and returns 
a survey and census dataset
}}

\references{
Guadarrama Sanz, M., Molina, I., and Rao, J.N.K.  (2016). \emph{A comparison of small 
 area estimation methods for poverty mapping}. In: 17 (Mar. 2016), 41-66 and 156 and 158.
}
