% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brazil.R
\docType{data}
\name{brazil}
\alias{brazil}
\title{Something}
\format{A data frame with 70139 rows and 44 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{}
}
\usage{
brazil
}
\description{
A dataset contain information about households in brazil
}
\details{
Courtesy to: Minnesota Population Center, 2017, Integrated Public Use Microdata Series, International: Version 6.5: Minneapolis, MN: University of Minnesota
}
\keyword{datasets}
