% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset-documentation.R
\docType{data}
\name{brazil}
\alias{brazil}
\title{Parts of the ?YEAR? census in Brazil}
\format{A data frame with 350696 rows and 11 variables:
\describe{
 \item{hh_inc}{household income}
 \item{geo2_br}{Consistent level two boundaries of Brazil, Municipality 1980 - 2010}
 \item{year}{year of the census}
 \item{age}{The age of a person in years}
 \item{sex}{sex as factor 2 and 1}
 \item{adults}{number of adults}
 \item{children}{number of children}
 \item{religion}{?}
 \item{computer}{factor 1 or 2}
 \item{trash}{?}
 \item{urban}{yes or no (2 or 1)}
}}
\source{
\url{https://international.ipums.org/international/index.shtml}
}
\usage{
data(brazil)
}
\description{
The dataset contains 350,696 individual observations from the ?YEAR? census in Brazil. 
Number of variables and observations are reduced to a size for a reasonable example 
of our ELLsae package
}
\keyword{datasets}
