% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Description-ELLsae.R, R/ELLsae.R
\docType{package}
\name{ELLsae}
\alias{ELLsae}
\alias{ELLsae-package}
\alias{sae,}
\alias{ELL,}
\alias{ellsae}
\title{ELLsae: A package for small area estimation}
\usage{
ELLsae(model, surveydata, censusdata, location_survey, mResponse,
  location_census, n_boot = 50, welfare.function, test, parallel = F)
}
\arguments{
\item{model}{a model that is specified for the relationship betwenn the
response varibale and the regressors. Model must be a linear model that can
be processed by \code{lm()}}

\item{surveydata}{Smaller surveydata with additional response variable of
interest. Will be used to estimate the linear model}

\item{censusdata}{The dataset in which a certain variable is supposed to be
imputed}

\item{location_survey}{Name of location variable or vector for the survey
data which is used for error correction and the location means (if
\code{mResponse} is specified)}

\item{mResponse}{Additional parameters for the regression based on location
means calculated from the census data to account for the lack of
information in a small survey}

\item{location_census}{name of location variable (string) in the census data
which is used for error correction and location means. If \code{mResponse}
is specified, but \code{location_census} is missing}

\item{n_boot}{Number of bootstrap samples used for the estimation, default is
\code{n_boot = 50}}

\item{welfare.function}{Additionally a welfare function for the response can
be specified}

\item{parallel}{indicates if compution is supposed to be done in parallel to
improve speed}
}
\value{
The function returns a vector of the imputed variable as well as ...
}
\description{
The method for small area estimation added by this package is largely based on an approach developed 
by C. Elbers, J. Lanjouw and P. Lanjouw (2003). It can be used to impute a missing variable 
from a smaller survey dataset into a census. The imputation is based on a 
linear model and bootstrap samples from a normal distribution. As of now the correction for heteroscedasticity 
is left omitted.

\code{ELLsae} is a method for small area estimation used to impute a missing
variable from a smaller survey dataset into a census. The imputation is based
on a linear model and bootstrap samples.
}
\details{
brief overview over the usage of the package
}
\section{ELLsae functions}{

Delivered are two functions that do about the same thing ...
}

\examples{
no examples are currently specified
}
\references{
\insertRef{sae1}{ELLsae}
 \insertRef{SAEcomparison}{ELLsae}
 Elbers, C., Lanjouw, J. O. and Lanjouw, P. (2003). \emph{Micro-Level Estimation of Poverty and Inequality}. 
  In: Econometrica 71.1, pp. 355-364, Jan 2003

A
}
\author{
Nikos Bosse [aut, cre], Felix Suettmann [aut, cre]
}
\keyword{SAE}
\keyword{imputation}
