% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELLsaeBig.R
\name{ELLsaeBig}
\alias{ELLsaeBig}
\title{ELLsaeBig}
\usage{
ELLsaeBig(model, surveydata, censusdata, location_survey, mResponse,
  location_census, n_boot = 50, welfare.function, test, parallel = F)
}
\arguments{
\item{model}{a model that is specified for the relationship betwenn
the response varibale and the regressors. Model must be a linear model that can be processed by \code{lm()}}

\item{surveydata}{Smaller surveydata with additional response variable of interest.
Will be used to estimate the linear model}

\item{censusdata}{The dataset in which a certain variable is supposed to be imputed}

\item{location_survey}{Name of location variable or vector for the survey data which is used for
error correction and the location means (if \code{mResponse} is specified)}

\item{mResponse}{Additional parameters for the regression based on location means
calculated from the census data to account for the lack of information in a small survey}

\item{n_boot}{Number of bootstrap samples used for the estimation, default is 50}

\item{welfare.function}{Additionally a welfare function for the response can be specified}
}
\value{
Was die Funktion ausspuckt.
}
\description{
Beschreibung der Funktion
}
